Company logos, almost fully responsive
Strange effect at around 900 pixels down to 600 where the img are stacked next to eachother but it goes outside the vw. I fixed this with overflow-z: hidden as a temp solution.

I been trying to do this for way to long now so I will be happy with it working for desctop and mobile then come back to the problem